Rename XDH to XDH_LEGACY
