+++ pgpainless-core/src/main/kotlin/org/pgpainless/key/generation/type/xdh_legacy/XDHLegacy.kt
@@ -2,19 +2,19 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package org.pgpainless.key.generation.type.xdh
+package org.pgpainless.key.generation.type.xdh_legacy
 
 import org.bouncycastle.jce.spec.ECNamedCurveGenParameterSpec
 import org.pgpainless.algorithm.PublicKeyAlgorithm
 import org.pgpainless.key.generation.type.KeyType
 
-class XDH private constructor(spec: XDHSpec) : KeyType {
+class XDHLegacy private constructor(spec: XDHLegacySpec) : KeyType {
     override val name = "XDH"
     override val algorithm = PublicKeyAlgorithm.ECDH
     override val bitStrength = spec.bitStrength
     override val algorithmSpec = ECNamedCurveGenParameterSpec(spec.algorithmName)
 
     companion object {
-        @JvmStatic fun fromSpec(spec: XDHSpec) = XDH(spec)
+        @JvmStatic fun fromSpec(spec: XDHLegacySpec) = XDHLegacy(spec)
     }
 }
